+++ azext_iot/constants.py
@@ -11,9 +11,24 @@ VERSION = "0.8.7"
 EXTENSION_NAME = "azure-cli-iot-ext"
 EXTENSION_ROOT = os.path.dirname(os.path.abspath(__file__))
 EXTENSION_CONFIG_ROOT_KEY = "iotext"
-EDGE_DEPLOYMENT_SCHEMA_2_PATH = os.path.join(EXTENSION_ROOT, 'assets', 'edge-deploy-2.0.schema.json')
+EDGE_DEPLOYMENT_SCHEMA_2_PATH = os.path.join(
+    EXTENSION_ROOT, "assets", "edge-deploy-2.0.schema.json"
+)
 BASE_API_VERSION = "2018-08-30-preview"
 BASE_MQTT_API_VERSION = "2018-06-30"
+MESSAGING_HTTP_C2D_SYSTEM_PROPERTIES = [
+    "iothub-messageid",
+    "iothub-correlationid",
+    "iothub-sequencenumber",
+    "iothub-to",
+    "iothub-userid",
+    "iothub-ack",
+    "iothub-expiry",
+    "iothub-deliverycount",
+    "iothub-enqueuedtime",
+    "ContentType",
+    "ContentEncoding",
+]
 METHOD_INVOKE_MAX_TIMEOUT_SEC = 300
 METHOD_INVOKE_MIN_TIMEOUT_SEC = 10
 MIN_SIM_MSG_INTERVAL = 1
@@ -29,7 +44,7 @@ TRACING_ALLOWED_FOR_SKU = "standard"
 USER_AGENT = "IoTPlatformCliExtension/{}".format(VERSION)
 
 # (Lib name, minimum version (including), maximum version (excluding))
-EVENT_LIB = ("uamqp", "1.0.3", "1.1")
+EVENT_LIB = ("uamqp", "1.1.0", "1.2.5")
 
 # Config Key's
 CONFIG_KEY_UAMQP_EXT_VERSION = "uamqp_ext_version"